requirements.txt
apiclient==1.0.4
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
Django==4.2
django-rest-authtoken==2.1.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
urllib3==1.26.15
